requirements.txt
langchain
flask
sentence-transformers
pypdf
python-dotenv
langchain-pinecone
langchain-community
langchain-core
langchain-google-genai
google-generativeai
langchain_huggingface
-e .